hacky solution to solve D73564127's duplicate dependency issues to unblock pybinding backend migration
Pull Request resolved: https://github.com/pytorch/executorch/pull/11566

This diff creates different namespace for Module and BundledModule under different mode (aten or portable) to hacky solve the duplicated symbol issues in D73564127 since the their are too many complex dependency issues to solve, and lots of users are asking for pybinding backend migration.

In the future this impl should be replaced by shim-like layers.
ghstack-source-id: 289776208

Differential Revision: [D76405701](https://our.internmc.facebook.com/intern/diff/D76405701/)